Conflicts:
	automation/Build.py
	automation/Common.py
	automation/Hudson_Common.py
	src/CMakeLists.txt
	src/backend/ocl_cpu_backend/CPUProgramBuilder.cpp
	src/backend/ocl_cpu_backend/ProgramBuilder.cpp

+++ opencl-intel/backend/validations/Comparator/IComparisonResults.h
@@ -286,7 +286,7 @@ namespace Validation {
             {
                 if((!ref.IsValid()) || (!act.IsValid()))
                     throw Exception::InvalidArgument("Input value container was invalid during comparator mismatch reporting");
-                double res = (T)abs((long)(ref.GetVal<T>()- act.GetVal<T>()));
+                double res = (T)abs((double)(long)(ref.GetVal<T>()- act.GetVal<T>()));
                 return res;
             }
